Move Time Range and User
