+++ src/main/webapp/components/time-range/index.js
@@ -1,0 +1,17 @@
+import TimeRange, {
+  TimeRangeAfter,
+  TimeRangeBefore,
+  TimeRangeDuring,
+  TimeRangeRelative,
+  validate,
+  createTimeRange,
+} from './time-range'
+export {
+  TimeRange as default,
+  TimeRangeAfter,
+  TimeRangeBefore,
+  TimeRangeDuring,
+  TimeRangeRelative,
+  validate,
+  createTimeRange,
+}
