+++ src/main/webapp/components/time-range/time-range.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { storiesOf } from './@storybook/react'
+import { storiesOf } from '../../@storybook/react'
 import { action } from '@connexta/ace/@storybook/addon-actions'
 import TimeRange, { validate } from './time-range'
