+++ src/main/webapp/components/user/index.js
@@ -1,0 +1,2 @@
+import UserGQL, { User } from './user'
+export { UserGQL as default, User }
